backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lessonscribe-backend"
version = "0.1.0"
description = "Local FastAPI backend for LessonScribe."
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "openai-whisper",
  "python-multipart",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
  "httpx",
  "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
